requirements.txt
RPi.GPIO>=0.7.1
flask>=2.3.0
smbus>=1.1.post2
picamera2>=0.3.12
opencv-python>=4.8.0
numpy>=1.24.0

